добавлена обработка исключений в эндпоинты

+++ webapp/api/education/lesson.py
@@ -24,9 +24,12 @@ async def create_lesson_endpoint(
     session: AsyncSession = Depends(get_session),
     current_user: JwtTokenT = Depends(jwt_auth.get_current_user),
 ):
-    if current_user['role'] == 'admin' or 'teacher':
+    if current_user['role'] not in ['admin', 'teacher']:
+        raise HTTPException(status_code=403, detail='Нет доступа для выполнения этой операции')
+    try:
         return await create_lesson(session=session, course_id=course_id, lesson_data=lesson_data)
-    raise HTTPException(status_code=403, detail='Нет доступа для выполнения этой операции')
+    except Exception as e:
+        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail=str(e))
 
 
 @lesson_router.get('/', response_model=List[LessonRead], tags=['Lessons'], response_class=ORJSONResponse)
@@ -35,7 +38,13 @@ async def get_all_lessons_by_course_endpoint(
     session: AsyncSession = Depends(get_session),
     current_user: JwtTokenT = Depends(jwt_auth.get_current_user),
 ):
-    return await get_lessons_all_by_course_id(session=session, course_id=course_id)
+    try:
+        lessons = await get_lessons_all_by_course_id(session=session, course_id=course_id)
+        if lessons is None:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Уроки не найдены')
+        return lessons
+    except Exception as e:
+        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail=str(e))
 
 
 @lesson_router.get('/{lesson_id}', response_model=LessonRead, tags=['Lessons'])
@@ -45,10 +54,13 @@ async def get_lesson_endpoint(
     session: AsyncSession = Depends(get_session),
     current_user: JwtTokenT = Depends(jwt_auth.get_current_user),
 ):
-    lesson = await get_lesson_by_id(session=session, course_id=course_id, lesson_id=lesson_id)
-    if lesson is None:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Lesson not found')
-    return lesson
+    try:
+        lesson = await get_lesson_by_id(session=session, course_id=course_id, lesson_id=lesson_id)
+        if lesson is None:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Урок не найден')
+        return lesson
+    except Exception as e:
+        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail=str(e))
 
 
 @lesson_router.put('/{lesson_id}', response_model=LessonRead, tags=['Lessons'])
@@ -59,12 +71,17 @@ async def update_lesson_endpoint(
     session: AsyncSession = Depends(get_session),
     current_user: JwtTokenT = Depends(jwt_auth.get_current_user),
 ):
-    updated_lesson = await update_lesson(
-        session=session, course_id=course_id, lesson_id=lesson_id, lesson_data=lesson_data
-    )
-    if updated_lesson is None:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Lesson not found')
-    return updated_lesson
+    if current_user['role'] not in ['admin', 'teacher']:
+        raise HTTPException(status_code=403, detail='Нет доступа для выполнения этой операции')
+    try:
+        updated_lesson = await update_lesson(
+            session=session, course_id=course_id, lesson_id=lesson_id, lesson_data=lesson_data
+        )
+        if updated_lesson is None:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Урок не найден')
+        return updated_lesson
+    except Exception as e:
+        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail=str(e))
 
 
 @lesson_router.delete('/{lesson_id}', status_code=status.HTTP_204_NO_CONTENT, tags=['Lessons'])
@@ -74,7 +91,11 @@ async def delete_lesson_endpoint(
     session: AsyncSession = Depends(get_session),
     current_user: JwtTokenT = Depends(jwt_auth.get_current_user),
 ):
+    if current_user['role'] not in ['admin', 'teacher']:
+        raise HTTPException(status_code=403, detail='Нет доступа для выполнения этой операции')
     try:
-        return await delete_lesson(session=session, course_id=course_id, lesson_id=lesson_id)
-    except HTTPException:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Lesson not found')
+        result = await delete_lesson(session=session, course_id=course_id, lesson_id=lesson_id)
+        if not result:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Урок не найден')
+    except Exception as e:
+        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail=str(e))
